refactor(kimi): separate interaction state validation and transport
